+++ datapool/DataSet.h
@@ -121,8 +121,8 @@ public:
 	 * @param bigEndian
 	 * @return
 	 */
-	ReturnValue_t serialize(uint8_t** buffer, uint32_t* size,
-			const uint32_t max_size, bool bigEndian) const;
+	ReturnValue_t serialize(uint8_t** buffer, size_t* size,
+			const size_t max_size, bool bigEndian) const;
 
 	uint32_t getSerializedSize() const;
