+++ datapool/PoolVariable.h
@@ -193,8 +193,8 @@ public:
 		return *this;
 	}
 
-	virtual ReturnValue_t serialize(uint8_t** buffer, uint32_t* size,
-			const uint32_t max_size, bool bigEndian) const {
+	virtual ReturnValue_t serialize(uint8_t** buffer, size_t* size,
+			const size_t max_size, bool bigEndian) const {
 		return SerializeAdapter<T>::serialize(&value, buffer, size, max_size,
 				bigEndian);
 	}
