+++ power/Fuse.h
@@ -49,8 +49,8 @@ public:
 	uint8_t getFuseId() const;
 	ReturnValue_t initialize();
 	DeviceList devices;
-	ReturnValue_t serialize(uint8_t** buffer, uint32_t* size,
-			const uint32_t max_size, bool bigEndian) const;
+	ReturnValue_t serialize(uint8_t** buffer, size_t* size,
+			const size_t max_size, bool bigEndian) const;
 	uint32_t getSerializedSize() const;
 	ReturnValue_t deSerialize(const uint8_t** buffer, int32_t* size,
 	bool bigEndian);
